python/requirements.txt
pkg-resources==0.0.0
pyusb==1.1.0
pyserial

